Upgrade project dependencies and improve code robustness
The project dependencies have been updated to their latest version, in light of new features and performance improvements offered by them. This includes an update to Kotlin's version to 1.9.0 along with the corresponding serialization plugin, Gradle to 8.3, and various dependencies in the build.gradle.kts file.

Changes were made to the Undertow server to get better control over the incoming HTTP requests, including setting the Content-type header for all responses and better logging of request redirects and not found errors. A new MetricsHandlerWrapper has also been added to provide more insights about the server's performance.

Funding option of the project (FUNDING.yml) has been removed as it's no longer needed.

The Gradle Wrapper scripts have been updated to match the new version of Gradle and provide better compatibility.

Docker-related changes include an update to the base image for the Dockerfile and build and push tooling in the GitHub Actions workflow file.

.gitattributes file has been added to handle line ending discrepancies across different platforms.